Create MD5 Hash Crack
This is a Python function that takes: a hashed message, a salt, and a word list, and attempts to decipher said hashed message. 
Simply replace the : hash, salt, and filename variables with the appropriate information. 
The s1 = salt.decode('hex') line is to check for a hex encoded line when checking against the hash variable "p".